+++ selene-vscode/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "selene-vscode" extension will be documented in this 
 
 If you want to stay up to date with selene itself, you can find the changelog in [selene's CHANGELOG.md](https://github.com/Kampfkarren/selene/blob/master/CHANGELOG.md).
 
-## Unreleased
+## [1.2.0]
 - Temporary paths, such as Git diff previews, will now use your current workspace folder for configuration.
 - Fixed a bug where the extension would become unusable if it couldn't connect to the internet.
 - If a Roblox codebase is detected without a standard library, you will be prompted to set it up. You can disable this in the settings if you work in non-Roblox codebases that trip them up.